requirements.txt
requests
tqdm
colorama
sentence_transformers
pandas
faiss-cpu
numpy